refactor: 랜딩페이지 로직 분리
- peer 설정 및 join 이벤트 리스너 분리
